[chore] Clear newly found warnings/errors
4189: 'scaleOver75': local variable is initialized but not referenced
4389: '==': signed/unsigned mismatch
4701: potentially uninitialized local variable
4702: unreachable code
4703: potentially uninitialized local pointer variable
4706: assignment within conditional expression

+++ src/map/ai/controllers/trust_controller.cpp
@@ -492,7 +492,7 @@ uint8 CTrustController::GetPartyPosition()
     {
         if (trustList.at(i)->id == POwner->id)
         {
-            return i;
+            return static_cast<uint8>(i);
         }
     }
     return 0;
